requirements.txt
fastapi
uvicorn
openai
faiss-cpu
beautifulsoup4
requests
python-dotenv
pydantic
numpy
